Renamed Mixer_SDL.h/.cpp to MixerSDL.h/.cpp.
Updated references in solution files and modified preprocessor includes.
