+++ proj/vs2019/NAS2D.vcxproj.filters
@@ -90,9 +90,6 @@
     <ClCompile Include="..\..\src\Renderer\Renderer.cpp">
       <Filter>Source Files\Renderer</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\Mixer\Mixer_SDL.cpp">
-      <Filter>Source Files\Mixer</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\Exception.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -140,6 +137,9 @@
     </ClCompile>
     <ClCompile Include="..\..\src\Renderer\RendererOpenGL.cpp">
       <Filter>Source Files\Renderer</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\Mixer\MixerSDL.cpp">
+      <Filter>Source Files\Mixer</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -224,9 +224,6 @@
     <ClInclude Include="..\..\include\NAS2D\Renderer\Renderer.h">
       <Filter>Header Files\Renderer</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\include\NAS2D\Mixer\Mixer_SDL.h">
-      <Filter>Header Files\Mixer</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\include\NAS2D\Resources\ImageInfo.h">
       <Filter>Header Files\Resources</Filter>
     </ClInclude>
@@ -280,6 +277,9 @@
     </ClInclude>
     <ClInclude Include="..\..\include\NAS2D\Renderer\RendererOpenGL.h">
       <Filter>Header Files\Renderer</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\include\NAS2D\Mixer\MixerSDL.h">
+      <Filter>Header Files\Mixer</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
